network-config-agent/requirements.txt
netmiko
openai
pyyaml
python-dotenv